requirements.txt
panda
numpy
altair
LogisticRegression
train_test_split
classification_report
confusion_matrix
metrics
matplotlib.pyplot